The component created in this program implements the IX and IY interfaces, as well as IUnknown, which is the base interface for all COM interfaces. To access interfaces, you must execute a QueryInterface query that returns a pointer to the required interface. Each bean must also implement the CreateInstance function, which instantiates the bean and returns a pointer to that bean's IUnknown.
The client application uses the CreateInstance function to instantiate the bean and obtain a pointer to the bean's IUnknown. The client application then queries the IX, IY, and IZ interfaces via IUnknown using the QueryInterface method. The results of the queries are used to call the member methods of the interfaces.